server_v2/requirements.txt
openai
langchain
jinja2
python-dotenv
PyPDF2
tika